make the "fetch all posts function" works
